+++ packages/server/graphql/intranetSchema/intranetSchema.ts
@@ -39,6 +39,7 @@ import pingActionTick from './queries/pingActionTick'
 import signups from './queries/signups'
 import user from './queries/user'
 import users from './queries/users'
+import updateEmail from './mutations/updateEmail'
 
 const query = new GraphQLObjectType<any, GQLContext>({
   name: 'Query',
@@ -88,7 +89,8 @@ const mutation = new GraphQLObjectType<any, GQLContext>({
       stripeSucceedPayment,
       stripeUpdateCreditCard,
       stripeUpdateInvoiceItem,
-      stripeInvoiceFinalized
+      stripeInvoiceFinalized,
+      updateEmail
     } as any)
 })