+++ packages/server/postgres/queries/src/updateUserQuery.sql
@@ -6,7 +6,7 @@ UPDATE "User" SET
   email = COALESCE(:email, "email"),
   "updatedAt" = COALESCE(:updatedAt, "updatedAt"),
   inactive = COALESCE(:inactive, "inactive"),
-  "lastSeenAt" = COALESCE(:lastSeenAt, "lastSeenAt"),
+  "lastSeenAt" = GREATEST("lastSeenAt", COALESCE(:lastSeenAt, "lastSeenAt")),
   "preferredName" = COALESCE(:preferredName, "preferredName"),
   tier = COALESCE(:tier, "tier"),
   picture = COALESCE(:picture, "picture"),